Starting Database + Images
To get the database working in XAMPP, simply copy/paste the contents of the file in the dump_file folder into the textbox at: http://localhost/phpmyadmin/index.php?route=/server/sql

After doing this press go, and the database will be built locally for you.

The images_skins file contains images of all of the skins in the database with names that relate to the "ImgID" section in the database. Each ImgID pertains to one of the images in the folder, the ".webp" tag will just need to be added on to access the actual image in the folder fyi.