starting database + images + dropdown search bar
Added a database dump file. To build your database just take the contents of the dump file and paste them into your xampp mysql control panel.

Added folder of skin images. Each database entry has a connected "ImgID" that is connected to the name of an image in the the folder (.webp must be added to the end of this ID to actually get the image though!)

Changed the .html file to a .php file so it can access the database.

Added skeleton of a search bar + dropdown menu that will contain all of the skins the player can guess.

Javascript file now contains logic for the above mentioned search function.

CSS file contains basic styling for the search function.